change the type of "power" variable

+++ 05-METHODS/08 Math Power/Program.cs
@@ -5,12 +5,12 @@ class Program
     static void Main()
     {
         double number = double.Parse(Console.ReadLine());
-        double power = int.Parse(Console.ReadLine());
+        int power = int.Parse(Console.ReadLine());
         double result = RaiseToPower(number, power);
         Console.WriteLine(result);
     }
 
-    private static double RaiseToPower(double number, double power)
+    private static double RaiseToPower(double number, int power)
     {
         return Math.Pow(number, power);
     }
